requirements.txt
python-dotenv
pandas
numpy
pymongo
certifi
scikit-learn
dill
pyaml
dagshub
mlflow<3
fastapi
uvicorn
#-e .